Decouples spritesheet and bitmap text middleware from loaders
