+++ packages/loaders/test/index.js
@@ -1,4 +1,2 @@
-require('./bitmapFontParser');
 require('./loader');
-require('./spritesheetParser');
 require('./textureParser');
