+++ packages/spritesheet/package.json
@@ -31,7 +31,8 @@
   "dependencies": {
     "@pixi/core": "^5.0.0-alpha",
     "@pixi/math": "^5.0.0-alpha",
-    "@pixi/utils": "^5.0.0-alpha"
+    "@pixi/utils": "^5.0.0-alpha",
+    "@pixi/loaders": "^5.0.0-alpha"
   },
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
