+++ packages/spritesheet/src/index.js
@@ -1,1 +1,2 @@
 export { default as Spritesheet } from './Spritesheet';
+export { default as SpritesheetLoader } from './SpritesheetLoader';
