+++ packages/text-bitmap/src/index.js
@@ -1,1 +1,2 @@
 export { default as BitmapText } from './BitmapText';
+export { default as BitmapFontLoader } from './BitmapFontLoader';
