+++ packages/text-bitmap/test/resources/atlas.json
