+++ packages/text-bitmap/test/resources/atlas.png
