+++ packages/text-bitmap/test/resources/atlas@0.5x.json
