+++ packages/text-bitmap/test/resources/font.fnt
