+++ packages/text-bitmap/test/resources/font.png
