+++ packages/text-bitmap/test/resources/font@0.5x.fnt
